% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variaciones.R
\name{Variaciones}
\alias{Variaciones}
\title{Variations}
\usage{
Variaciones(sp, r, mc, mp, z, y, x)
}
\arguments{
\item{sp}{Species number of the simulator}

\item{r}{Number of resources in the simulator}

\item{mc}{Media of consumption of the species}

\item{mp}{Media of production of the species}

\item{z}{The parameter that will change in next iteration, 1= sp, 2= r, 3= mc, 4= mp.}

\item{y}{Value that will be add in each iteration}

\item{x}{Number of iterations}
}
\value{
A data frame that contains the A,a, b values of the different Rank-Abundance curves
}
\description{
The Variations function provides a visualization about the effect of different parameters on the Rank-abundance curve.
}
\author{
Erik Martinez
}
